Refactor test files, get load tests working
