+++ cumulusci/tasks/bulkdata/tests/test_load.py
@@ -1,0 +1,830 @@
+import os
+import json
+import responses
+import shutil
+import unittest
+
+from collections import OrderedDict
+from datetime import datetime
+from sqlalchemy import Column, Table, Unicode
+from unittest import mock
+
+from cumulusci.utils import temporary_dir
+from cumulusci.core.exceptions import BulkDataException, TaskOptionsError
+from cumulusci.tasks.bulkdata import LoadData
+from cumulusci.tasks.bulkdata.step import (
+    Result,
+    Operation,
+    Status,
+    DmlStep,
+    BulkApiDmlStep,
+    BulkApiQueryStep,
+)
+from cumulusci.tasks.bulkdata.tests.utils import _make_task
+
+BULK_BATCH_RESPONSE = '<root xmlns="http://ns"><batch><state>{}</state></batch></root>'
+
+MAPPING_FILE = """Insert Households:
+    sf_object: Account
+    table: households
+    fields:
+        Id: sf_id
+    static:
+        Name: TestHousehold
+    record_type: HH_Account
+Insert Contacts:
+    sf_object: Contact
+    table: contacts
+    filters:
+        - 'household_id is not null'
+    fields:
+        Id: sf_id
+        FirstName: first_name
+        LastName: last_name
+        Email: email
+    lookups:
+        AccountId:
+            key_field: household_id
+            table: households
+            join_field: household_id
+            value_field: sf_id
+"""
+
+
+class MockBulkApiDmlStep(DmlStep):
+    def __init__(
+        self, sobject: str, operation: Operation, api_options: dict, context, fields
+    ):
+        super().__init__(sobject, operation, api_options, context, fields)
+        self.results = []
+        self.records = []
+
+    def start(self):
+        self.job_id = "JOB"
+
+    def end(self):
+        self.status = Status.SUCCESS
+
+    def load_records(self, records):
+        self.records.extend(records)
+
+    def get_results(self):
+        return iter(self.results)
+
+
+class test_LoadData(unittest.TestCase):
+    mapping_file = "mapping_v1.yml"
+
+    @responses.activate
+    @mock.patch("cumulusci.tasks.bulkdata.load.BulkApiDmlStep")
+    def test_run(self, step_mock):
+        responses.add(
+            method="GET",
+            url="https://example.com/services/data/vNone/query/?q=SELECT+Id+FROM+RecordType+WHERE+SObjectType%3D%27Account%27AND+DeveloperName+%3D+%27HH_Account%27+LIMIT+1",
+            body=json.dumps({"records": [{"Id": "1"}]}),
+            status=200,
+        )
+
+        base_path = os.path.dirname(__file__)
+        db_path = os.path.join(base_path, "testdata.db")
+        mapping_path = os.path.join(base_path, self.mapping_file)
+
+        with temporary_dir() as d:
+            tmp_db_path = os.path.join(d, "testdata.db")
+            shutil.copyfile(db_path, tmp_db_path)
+
+            task = _make_task(
+                LoadData,
+                {
+                    "options": {
+                        "database_url": f"sqlite:///{tmp_db_path}",
+                        "mapping": mapping_path,
+                    }
+                },
+            )
+
+            task.bulk = mock.Mock()
+            task.sf = mock.Mock()
+
+            step = MockBulkApiDmlStep("Contact", Operation.INSERT, {}, task, [])
+            step_mock.return_value = step
+
+            step.results = [
+                Result("001000000000000", True, None),
+                Result("003000000000000", True, None),
+                Result("003000000000001", True, None),
+            ]
+
+            task()
+
+        assert step.records == [
+            ["TestHousehold", "1"],
+            ["Test", "User", "test@example.com", "001000000000000"],
+            ["Error", "User", "error@example.com", "001000000000000"],
+        ]
+
+        # FIXME: how do we assert that the Ids were updated in the DB?
+
+    def test_run_task__start_step(self):
+        task = _make_task(
+            LoadData,
+            {
+                "options": {
+                    "database_url": "sqlite://",
+                    "mapping": "mapping.yml",
+                    "start_step": "Insert Contacts",
+                }
+            },
+        )
+        task._init_db = mock.Mock()
+        task._init_mapping = mock.Mock()
+        task.mapping = {}
+        task.mapping["Insert Households"] = {"one": 1}
+        task.mapping["Insert Contacts"] = {"two": 2}
+        task.after_steps = {}
+        task._load_mapping = mock.Mock(return_value="Completed")
+        task()
+        task._load_mapping.assert_called_once_with({"two": 2, "action": "insert"})
+
+    def test_run_task__after_steps(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        task._init_db = mock.Mock()
+        task._init_mapping = mock.Mock()
+        task._expand_mapping = mock.Mock()
+        task.mapping = {}
+        task.mapping["Insert Households"] = 1
+        task.mapping["Insert Contacts"] = 2
+        households_steps = {}
+        households_steps["four"] = 4
+        households_steps["five"] = 5
+        task.after_steps = {
+            "Insert Contacts": {"three": 3},
+            "Insert Households": households_steps,
+        }
+        task._load_mapping = mock.Mock(return_value="Completed")
+        task()
+        task._load_mapping.assert_has_calls(
+            [mock.call(1), mock.call(4), mock.call(5), mock.call(2), mock.call(3)]
+        )
+
+    def test_run_task__after_steps_failure(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        task._init_db = mock.Mock()
+        task._init_mapping = mock.Mock()
+        task._expand_mapping = mock.Mock()
+        task.mapping = {}
+        task.mapping["Insert Households"] = 1
+        task.mapping["Insert Contacts"] = 2
+        households_steps = {}
+        households_steps["four"] = 4
+        households_steps["five"] = 5
+        task.after_steps = {
+            "Insert Contacts": {"three": 3},
+            "Insert Households": households_steps,
+        }
+        task._load_mapping = mock.Mock(side_effect=[Status.SUCCESS, Status.FAILURE])
+        with self.assertRaises(BulkDataException):
+            task()
+
+    @responses.activate
+    @mock.patch("cumulusci.tasks.bulkdata.load.BulkApiDmlStep")
+    def test_run__sql(self, step_mock):
+        responses.add(
+            method="GET",
+            url="https://example.com/services/data/vNone/query/?q=SELECT+Id+FROM+RecordType+WHERE+SObjectType%3D%27Account%27AND+DeveloperName+%3D+%27HH_Account%27+LIMIT+1",
+            body=json.dumps({"records": [{"Id": "1"}]}),
+            status=200,
+        )
+
+        base_path = os.path.dirname(__file__)
+        sql_path = os.path.join(base_path, "testdata.sql")
+        mapping_path = os.path.join(base_path, self.mapping_file)
+
+        task = _make_task(
+            LoadData, {"options": {"sql_path": sql_path, "mapping": mapping_path}}
+        )
+        task.bulk = mock.Mock()
+        task.sf = mock.Mock()
+        step = MockBulkApiDmlStep("Contact", Operation.INSERT, {}, task, [])
+        step_mock.return_value = step
+        step.results = [
+            Result("001000000000000", True, None),
+            Result("003000000000000", True, None),
+            Result("003000000000001", True, None),
+        ]
+        task()
+
+        assert step.records == [
+            ["TestHousehold", "1"],
+            ["Test", "User", "test@example.com", "001000000000000"],
+            ["Error", "User", "error@example.com", "001000000000000"],
+        ]
+
+        # FIXME: how do we assert that the Ids were updated in the DB?
+
+    def test_init_options__missing_input(self):
+        with self.assertRaises(TaskOptionsError):
+            _make_task(LoadData, {"options": {}})
+
+    def test_init_options__database_url(self):
+        t = _make_task(
+            LoadData,
+            {"options": {"database_url": "file:///test.db", "mapping": "mapping.yml"}},
+        )
+
+        assert t.options["database_url"] == "file:///test.db"
+        assert t.options["sql_path"] is None
+
+    def test_init_options__sql_path(self):
+        t = _make_task(
+            LoadData, {"options": {"sql_path": "test.sql", "mapping": "mapping.yml"}}
+        )
+
+        assert t.options["sql_path"] == "test.sql"
+        assert t.options["database_url"] is None
+
+    def test_expand_mapping_creates_after_steps(self):
+        base_path = os.path.dirname(__file__)
+        mapping_path = os.path.join(base_path, "mapping_after.yml")
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": mapping_path}},
+        )
+
+        task._init_mapping()
+
+        model = mock.Mock()
+        model.__table__ = mock.Mock()
+        model.__table__.primary_key.columns.keys.return_value = ["sf_id"]
+        task.models = {"accounts": model, "contacts": model}
+
+        task._expand_mapping()
+
+        self.assertEqual({}, task.after_steps["Insert Opportunities"])
+        self.assertEqual(
+            [
+                "Update Account Dependencies After Insert Contacts",
+                "Update Contact Dependencies After Insert Contacts",
+            ],
+            list(task.after_steps["Insert Contacts"].keys()),
+        )
+        lookups = {}
+        lookups["Id"] = {"table": "accounts", "key_field": "sf_id"}
+        lookups["Primary_Contact__c"] = {"table": "contacts"}
+        self.assertEqual(
+            {
+                "sf_object": "Account",
+                "action": "update",
+                "table": "accounts",
+                "lookups": lookups,
+                "fields": {},
+            },
+            task.after_steps["Insert Contacts"][
+                "Update Account Dependencies After Insert Contacts"
+            ],
+        )
+        lookups = {}
+        lookups["Id"] = {"table": "contacts", "key_field": "sf_id"}
+        lookups["ReportsToId"] = {"table": "contacts"}
+        self.assertEqual(
+            {
+                "sf_object": "Contact",
+                "action": "update",
+                "table": "contacts",
+                "fields": {},
+                "lookups": lookups,
+            },
+            task.after_steps["Insert Contacts"][
+                "Update Contact Dependencies After Insert Contacts"
+            ],
+        )
+        self.assertEqual(
+            ["Update Account Dependencies After Insert Accounts"],
+            list(task.after_steps["Insert Accounts"].keys()),
+        )
+        lookups = {}
+        lookups["Id"] = {"table": "accounts", "key_field": "sf_id"}
+        lookups["ParentId"] = {"table": "accounts"}
+        self.assertEqual(
+            {
+                "sf_object": "Account",
+                "action": "update",
+                "table": "accounts",
+                "fields": {},
+                "lookups": lookups,
+            },
+            task.after_steps["Insert Accounts"][
+                "Update Account Dependencies After Insert Accounts"
+            ],
+        )
+
+    # FIXME: rebuild
+    def test_stream_queried_data__skips_empty_rows(self):
+        task = _make_task(
+            LoadData, {"options": {"database_url": "sqlite://", "mapping": "test.yml"}}
+        )
+        task.sf = mock.Mock()
+
+        mapping = {
+            "sf_object": "Account",
+            "action": "update",
+            "fields": {},
+            "lookups": {
+                "Id": {"table": "accounts", "key_field": "sf_id"},
+                "ParentId": {"table": "accounts"},
+            },
+        }
+
+        task._query_db = mock.Mock()
+        task._query_db.return_value.yield_per = mock.Mock(
+            return_value=[
+                # Local Id, Loaded Id, Parent Id
+                ["001000000001", "001000000005", "001000000007"],
+                ["001000000002", "001000000006", "001000000008"],
+                ["001000000003", "001000000009", None],
+            ]
+        )
+
+        records = list(task._stream_queried_data(mapping))
+        self.assertEqual(
+            [["001000000005", "001000000007"], ["001000000006", "001000000008"]],
+            records,
+        )
+
+    def test_get_columns(self):
+        task = _make_task(
+            LoadData, {"options": {"database_url": "sqlite://", "mapping": "test.yml"}}
+        )
+
+        fields = {}
+        fields["Id"] = "sf_id"
+        fields["Name"] = "Name"
+
+        self.assertEqual(
+            ["Name", "Industry", "RecordTypeId"],
+            task._get_columns(
+                {
+                    "sf_object": "Account",
+                    "action": "insert",
+                    "fields": fields,
+                    "static": {"Industry": "Technology"},
+                    "record_type": "Organization",
+                }
+            ),
+        )
+        self.assertEqual(
+            ["Id", "Name", "Industry", "RecordTypeId"],
+            task._get_columns(
+                {
+                    "sf_object": "Account",
+                    "action": "update",
+                    "fields": fields,
+                    "static": {"Industry": "Technology"},
+                    "record_type": "Organization",
+                }
+            ),
+        )
+
+        fields["RecordTypeId"] = "recordtypeid"
+        fields["AccountSite"] = "accountsite"
+
+        self.assertEqual(
+            ["Id", "Name", "AccountSite", "Industry", "RecordTypeId"],
+            task._get_columns(
+                {
+                    "sf_object": "Account",
+                    "action": "update",
+                    "fields": fields,
+                    "static": {"Industry": "Technology"},
+                }
+            ),
+        )
+
+    def test_get_statics(self):
+        task = _make_task(
+            LoadData, {"options": {"database_url": "sqlite://", "mapping": "test.yml"}}
+        )
+        task.sf = mock.Mock()
+        task.sf.query.return_value = {"records": [{"Id": "012000000000000"}]}
+
+        self.assertEqual(
+            ["Technology", "012000000000000"],
+            task._get_statics(
+                {
+                    "sf_object": "Account",
+                    "action": "insert",
+                    "fields": {"Id": "sf_id", "Name": "Name"},
+                    "static": {"Industry": "Technology"},
+                    "record_type": "Organization",
+                }
+            ),
+        )
+
+    @mock.patch("cumulusci.tasks.bulkdata.load.aliased")
+    def test_query_db__joins_self_lookups(self, aliased):
+        task = _make_task(
+            LoadData, {"options": {"database_url": "sqlite://", "mapping": "test.yml"}}
+        )
+        model = mock.Mock()
+        task.models = {"accounts": model}
+        task.metadata = mock.Mock()
+        task.metadata.tables = {"accounts_sf_ids": mock.Mock()}
+        task.session = mock.Mock()
+
+        model.__table__ = mock.Mock()
+        model.__table__.primary_key.columns.keys.return_value = ["sf_id"]
+        columns = {"sf_id": mock.Mock(), "name": mock.Mock()}
+        model.__table__.columns = columns
+
+        mapping = {
+            "sf_object": "Account",
+            "table": "accounts",
+            "action": "update",
+            "oid_as_pk": True,
+            "fields": {"Id": "sf_id", "Name": "name"},
+            "lookups": {"ParentId": {"table": "accounts", "key_field": "sf_id"}},
+        }
+
+        task._query_db(mapping)
+
+        # Validate that the column set is accurate
+        task.session.query.assert_called_once_with(
+            model.sf_id,
+            model.__table__.columns["name"],
+            aliased.return_value.columns.sf_id,
+        )
+
+        # Validate that we asked for an outer join on the self-lookup
+        aliased.assert_called_once_with(task.metadata.tables["accounts_sf_ids"])
+        task.session.query.return_value.outerjoin.assert_called_once_with(
+            aliased.return_value, False
+        )
+
+    def test_convert(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        self.assertIsInstance(task._convert(datetime.now()), str)
+
+    def test_initialize_id_table__already_exists(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        task.mapping = {}
+        task._init_db()
+        id_table = Table(
+            "test_sf_ids", task.metadata, Column("id", Unicode(255), primary_key=True)
+        )
+        id_table.create()
+        task._initialize_id_table({"table": "test"}, True)
+        new_id_table = task.metadata.tables["test_sf_ids"]
+        self.assertFalse(new_id_table is id_table)
+
+    def test_initialize_id_table__already_exists_and_should_not_reset_table(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        task.mapping = {}
+        task._init_db()
+        id_table = Table(
+            "test_sf_ids", task.metadata, Column("id", Unicode(255), primary_key=True)
+        )
+        id_table.create()
+        table_name = task._initialize_id_table({"table": "test"}, False)
+        assert table_name == "test_sf_ids"
+        new_id_table = task.metadata.tables["test_sf_ids"]
+        assert new_id_table is id_table
+
+    def test_run_task__exception_failure(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+        task._init_db = mock.Mock()
+        task._init_mapping = mock.Mock()
+        task._load_mapping = mock.Mock(return_value=Status.FAILURE)
+        task.mapping = {"Test": {"test": "test"}}
+
+        with self.assertRaises(BulkDataException):
+            task()
+
+    def test_process_job_results__insert_success(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        task.session = mock.Mock()
+        task._initialize_id_table = mock.Mock()
+        task._sql_bulk_insert_from_records = mock.Mock()
+        task.bulk = mock.Mock()
+        task.sf = mock.Mock()
+        task.local_ids = ["1"]
+
+        step = MockBulkApiDmlStep("Contact", Operation.INSERT, {}, task, [])
+        step.results = [Result("001111111111111", True, None)]
+
+        mapping = {"table": "Account", "action": "insert"}
+        task._process_job_results(mapping, step)
+
+        task.session.connection.assert_called_once()
+        task._initialize_id_table.assert_called_once_with(mapping, True)
+        task._sql_bulk_insert_from_records.assert_called_once()
+        task.session.commit.assert_called_once()
+
+    def test_process_job_results__update_success(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        task.session = mock.Mock()
+        task._initialize_id_table = mock.Mock()
+        task._sql_bulk_insert_from_records = mock.Mock()
+        task.bulk = mock.Mock()
+        task.sf = mock.Mock()
+        task.local_ids = ["1"]
+
+        step = MockBulkApiDmlStep("Contact", Operation.INSERT, {}, task, [])
+        step.results = [Result("001111111111111", True, None)]
+
+        mapping = {"table": "Account", "action": "update"}
+        task._process_job_results(mapping, step)
+
+        task.session.connection.assert_not_called()
+        task._initialize_id_table.assert_not_called()
+        task._sql_bulk_insert_from_records.assert_not_called()
+        task.session.commit.assert_not_called()
+
+    def test_process_job_results__exception_failure(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        task.session = mock.Mock()
+        task._initialize_id_table = mock.Mock()
+        task._sql_bulk_insert_from_records = mock.Mock()
+        task.bulk = mock.Mock()
+        task.sf = mock.Mock()
+        task.local_ids = ["1"]
+
+        step = MockBulkApiDmlStep("Contact", Operation.UPDATE, {}, task, [])
+        step.results = [Result(None, False, "message")]
+
+        mapping = {"table": "Account", "action": "update"}
+
+        with self.assertRaises(BulkDataException) as ex:
+            task._process_job_results(mapping, step)
+
+        self.assertIn("Error on record with id", str(ex.exception))
+        self.assertIn("message", str(ex.exception))
+
+    def test_generate_results_id_map__success(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        step = mock.Mock()
+        step.get_results.return_value = iter(
+            [
+                Result("001000000000000", True, None),
+                Result("001000000000001", True, None),
+                Result("001000000000002", True, None),
+            ]
+        )
+
+        generator = task._generate_results_id_map(
+            step, ["001000000000009", "001000000000010", "001000000000011"]
+        )
+
+        assert list(generator) == [
+            ("001000000000009", "001000000000000"),
+            ("001000000000010", "001000000000001"),
+            ("001000000000011", "001000000000002"),
+        ]
+
+    def test_generate_results_id_map__exception_failure(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        step = mock.Mock()
+        step.get_results.return_value = iter(
+            [
+                Result("001000000000000", True, None),
+                Result(None, False, "error"),
+                Result("001000000000002", True, None),
+            ]
+        )
+
+        with self.assertRaises(BulkDataException) as ex:
+            list(
+                task._generate_results_id_map(
+                    step, ["001000000000009", "001000000000010", "001000000000011"]
+                )
+            )
+
+        self.assertIn("Error on record", str(ex.exception))
+        self.assertIn("001000000000010", str(ex.exception))
+
+    def test_generate_results_id_map__respects_silent_error_flag(self):
+        task = _make_task(
+            LoadData,
+            {
+                "options": {
+                    "ignore_row_errors": True,
+                    "database_url": "sqlite://",
+                    "mapping": "mapping.yml",
+                }
+            },
+        )
+
+        step = mock.Mock()
+        step.get_results.return_value = iter(
+            [
+                Result("001000000000000", True, None),
+                Result(None, False, None),
+                Result("001000000000002", True, None),
+            ]
+        )
+
+        generator = task._generate_results_id_map(
+            step, ["001000000000009", "001000000000010", "001000000000011"]
+        )
+
+        assert list(generator) == [
+            ("001000000000009", "001000000000000"),
+            ("001000000000011", "001000000000002"),
+        ]
+
+    @mock.patch("cumulusci.tasks.bulkdata.load.BulkApiDmlStep")
+    def test_load_mapping__record_type_mapping(self, step_mock):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        task.session = mock.Mock()
+        task._load_record_types = mock.Mock()
+        task._process_job_results = mock.Mock()
+
+        task._load_mapping(
+            {"sf_object": "Account", "action": "insert", "fields": {"Name": "Name"}}
+        )
+
+        task._load_record_types.assert_not_called()
+
+        task._load_mapping(
+            {
+                "sf_object": "Account",
+                "action": "insert",
+                "fields": {"Name": "Name", "RecordTypeId": "RecordTypeId"},
+            }
+        )
+        task._load_record_types.assert_called_once_with(
+            ["Account"], task.session.connection.return_value
+        )
+
+    def test_query_db__record_type_mapping(self):
+        task = _make_task(
+            LoadData, {"options": {"database_url": "sqlite://", "mapping": "test.yml"}}
+        )
+        model = mock.Mock()
+        task.models = {"accounts": model}
+        task.metadata = mock.Mock()
+        task.metadata.tables = {
+            "Account_rt_target_mapping": mock.Mock(),
+            "Account_rt_mapping": mock.Mock(),
+        }
+        task.session = mock.Mock()
+
+        model.__table__ = mock.Mock()
+        model.__table__.primary_key.columns.keys.return_value = ["sf_id"]
+        columns = {"sf_id": mock.Mock(), "name": mock.Mock()}
+        model.__table__.columns = columns
+
+        mapping = OrderedDict(
+            sf_object="Account",
+            table="accounts",
+            action="insert",
+            oid_as_pk=True,
+            fields=OrderedDict(Id="sf_id", Name="name", RecordTypeId="RecordTypeId"),
+        )
+
+        task._query_db(mapping)
+
+        # Validate that the column set is accurate
+        task.session.query.assert_called_once_with(
+            model.sf_id,
+            model.__table__.columns["name"],
+            task.metadata.tables["Account_rt_target_mapping"].columns.record_type_id,
+        )
+
+        # Validate that we asked for the right joins on the record type tables
+        task.session.query.return_value.outerjoin.assert_called_once_with(
+            task.metadata.tables["Account_rt_mapping"], False
+        )
+        task.session.query.return_value.outerjoin.return_value.outerjoin.assert_called_once_with(
+            task.metadata.tables["Account_rt_target_mapping"], False
+        )
+
+    # FIXME: break file link
+    @mock.patch("cumulusci.tasks.bulkdata.load.automap_base")
+    def test_init_db__record_type_mapping(self, base):
+        base_path = os.path.dirname(__file__)
+        mapping_path = os.path.join(base_path, self.mapping_file)
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": mapping_path}},
+        )
+
+        def create_table_mock(table_name):
+            task.models[table_name] = mock.Mock()
+
+        task._create_record_type_table = mock.Mock(side_effect=create_table_mock)
+        task.models = mock.Mock()
+        task.metadata = mock.Mock()
+
+        task._init_mapping()
+        task.mapping["Insert Households"]["fields"]["RecordTypeId"] = "RecordTypeId"
+        task._init_db()
+        task._create_record_type_table.assert_called_once_with(
+            "Account_rt_target_mapping"
+        )
+
+    def test_load_record_types(self):
+        task = _make_task(
+            LoadData,
+            {"options": {"database_url": "sqlite://", "mapping": "mapping.yml"}},
+        )
+
+        conn = mock.Mock()
+        task._extract_record_types = mock.Mock()
+        task._load_record_types(["Account", "Contact"], conn)
+        task._extract_record_types.assert_has_calls(
+            [
+                unittest.mock.call("Account", "Account_rt_target_mapping", conn),
+                unittest.mock.call("Contact", "Contact_rt_target_mapping", conn),
+            ]
+        )
+
+    @responses.activate
+    @mock.patch("cumulusci.tasks.bulkdata.load.BulkApiDmlStep")
+    def test_run__autopk(self, step_mock):
+        responses.add(
+            method="GET",
+            url="https://example.com/services/data/vNone/query/?q=SELECT+Id+FROM+RecordType+WHERE+SObjectType%3D%27Account%27AND+DeveloperName+%3D+%27HH_Account%27+LIMIT+1",
+            body=json.dumps({"records": [{"Id": "1"}]}),
+            status=200,
+        )
+
+        mapping_file = "mapping_v2.yml"
+        base_path = os.path.dirname(__file__)
+        db_path = os.path.join(base_path, "testdata.db")
+        mapping_path = os.path.join(base_path, mapping_file)
+        with temporary_dir() as d:
+            tmp_db_path = os.path.join(d, "testdata.db")
+            shutil.copyfile(db_path, tmp_db_path)
+
+            task = _make_task(
+                LoadData,
+                {
+                    "options": {
+                        "database_url": f"sqlite:///{tmp_db_path}",
+                        "mapping": mapping_path,
+                    }
+                },
+            )
+            task.bulk = mock.Mock()
+            task.sf = mock.Mock()
+            step = MockBulkApiDmlStep("Contact", Operation.INSERT, {}, task, [])
+            step_mock.return_value = step
+
+            step.results = [
+                Result("001000000000000", True, None),
+                Result("003000000000000", True, None),
+                Result("003000000000001", True, None),
+            ]
+
+            task()
+            task.session.close()
+
+        assert step.records == [
+            ["TestHousehold", "1"],
+            ["Test", "User", "test@example.com", "001000000000000"],
+            ["Error", "User", "error@example.com", "001000000000000"],
+        ]
+
+        # FIXME: how do we assert that the Ids were updated in the DB?
